Maya/229065 genericize table (#1587)
* warning comment for old table

* pull out generic table functionality

* move virtual bumper

* fix phylorun api mapping

* use generic table component for trees table

* begin a new home for column definition files

* types
