+++ src/frontend/src/common/queries/phyloRuns.ts
@@ -41,20 +41,10 @@ export const USE_PHYLO_RUN_INFO = {
   id: "phyloRunInfo",
 };
 
-/**
- * custom hook to automatically expire tree info when needed
- * such as when trees are deleted
- */
-export function usePhyloRunInfo(): UseQueryResult<PhyloRunResponse, unknown> {
+export function usePhyloRunInfo(): UseQueryResult<IdMap<PhyloRun>, unknown> {
   const state = store.getState();
   const pathogen = selectCurrentPathogen(state);
   return useQuery([USE_PHYLO_RUN_INFO, pathogen], fetchPhyloRuns, {
-    retry: false,
-  });
-}
-
-export function useNewPhyloRunInfo(): UseQueryResult<IdMap<PhyloRun>, unknown> {
-  return useQuery([USE_PHYLO_RUN_INFO], fetchPhyloRuns, {
     retry: false,
     select: mapPhyloRuns,
   });
