+++ src/frontend/src/common/utils/dataTransforms.ts
@@ -1,5 +1,6 @@
 import { generateOrgSpecificUrl, ORG_API } from "src/common/api";
 import ENV from "src/common/constants/ENV";
+import { TreeType } from "../constants/types";
 
 const { API_URL } = ENV;
 
@@ -64,10 +65,11 @@ export const getDownloadLinks = (phyloRun: PhyloRun): PhyloRunLinks => {
 /**
  * Converts a completely uppercase tree type to a capitalcase tree type.
  */
+type PhyloTreeManipulationType = PhyloRun & { tree_type?: TreeType };
 export const getCapitalCaseTreeType = (
-  phyloRun: PhyloRun
+  phyloRun: PhyloTreeManipulationType
 ): string | undefined => {
-  const { treeType } = phyloRun;
+  const { tree_type: treeType } = phyloRun;
 
   if (typeof treeType !== "string" || treeType.toLowerCase() == "unknown") {
     return undefined;
