+++ src/frontend/src/components/Table/style.ts
@@ -1,6 +1,16 @@
 import styled from "@emotion/styled";
 import { getColors, TableRow } from "czifui";
 
+// needed to keep search bar sticky
+export const StyledWrapper = styled.div`
+  flex: 1 1 auto;
+  overflow-y: auto;
+
+  & > div {
+    overflow: auto;
+  }
+`;
+
 export const StyledTableRow = styled(TableRow)`
   ${(props) => {
     const colors = getColors(props);
