+++ src/frontend/src/views/Data/components/DataNavigation/index.tsx
@@ -2,7 +2,7 @@ import { Tab } from "czifui";
 import { useRouter } from "next/router";
 import { useEffect, useState } from "react";
 import { useSelector } from "react-redux";
-import { useNewPhyloRunInfo as usePhyloRunInfo } from "src/common/queries/phyloRuns";
+import { usePhyloRunInfo } from "src/common/queries/phyloRuns";
 import { useNewSampleInfo as useSampleInfo } from "src/common/queries/samples";
 import { selectCurrentPathogen } from "src/common/redux/selectors";
 import { Pathogen } from "src/common/redux/types";
