+++ src/frontend/src/views/Data/components/SamplesView/components/SamplesTable/style.ts
@@ -1,12 +1,5 @@
 import styled from "@emotion/styled";
-import {
-  CellBasic,
-  getColors,
-  getFontWeights,
-  getSpaces,
-  InputCheckbox,
-  TableRow,
-} from "czifui";
+import { CellBasic, getFontWeights } from "czifui";
 
 export const StyledPrivateId = styled(CellBasic)`
   padding-left: 0;
@@ -32,34 +25,4 @@ export const StyledCellBasic = styled(CellBasic)`
       }
     `;
   }}
-`;
-
-export const StyledTableRow = styled(TableRow)`
-  ${(props) => {
-    const colors = getColors(props);
-    return `
-      &:hover {
-        background-color: ${colors?.primary[100]};
-      }
-    `;
-  }}
-`;
-
-export const StyledInputCheckbox = styled(InputCheckbox)`
-  ${(props) => {
-    const spaces = getSpaces(props);
-    return `
-      padding-bottom: ${spaces?.xxs}px;
-    `;
-  }}
-`;
-
-// needed to keep search bar sticky
-export const StyledWrapper = styled.div`
-  flex: 1 1 auto;
-  overflow-y: auto;
-
-  & > div {
-    overflow: auto;
-  }
 `;
