+++ src/frontend/src/views/Data/components/TreesView/index.tsx
@@ -1,12 +1,12 @@
 import { useState } from "react";
 import { HeadAppTitle } from "src/common/components";
-import { useNewPhyloRunInfo as usePhyloRunInfo } from "src/common/queries/phyloRuns";
+import { usePhyloRunInfo } from "src/common/queries/phyloRuns";
 import { IdMap } from "src/common/utils/dataTransforms";
 import { SearchBar } from "src/components/Table/components/SearchBar";
 import { DataNavigation } from "../DataNavigation";
 import { MaxWidth } from "../SamplesView/style";
 import { TreeCreateHelpLink } from "./components/TreeCreateHelpLink";
-import TreesTable from "./components/TreesTable";
+import { TreesTable } from "./components/TreesTable";
 import { Flex } from "./style";
 
 const TreesView = (): JSX.Element => {
